requirements.txt
asgiref==3.4.1
Django==3.2.9
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.4.0
psycopg2-binary==2.9.2
pydotplus==2.0.2
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
sqlparse==0.4.2
